requirements.txt
Django==3.0.3
django-heroku
django-jasmine
jasmine
gym[atari]
channels
numpy
hypothesis==4.38.0
attrs>=19.2
pytest-django
pickle-mixin
opencv-python
opencv-contrib-python
opencv-python-headless
